+++ src/main/java/chess/webdto/PieceInfosDto.java
@@ -1,0 +1,6 @@
+package chess.webdto;
+
+import java.util.List;
+
+public class PieceInfosDto {
+}
