--- src/main/java/chess/webdto/PositionAndPieceDto.java
@@ -1,4 +1,0 @@
-package chess.webdto;
-
-public class PositionAndPieceDto {
-}
